modified warings in mac

+++ tlang/compiler/src/tl-compile-option.cpp
@@ -82,8 +82,7 @@ CompileOption::getOneOption(const char* argv[], int &leftArgc, string &key, stri
 }
 
 /// iterate options map table, find key and val
-void 
-CompileOption::parseAllOptions(map<string, string> &options)
+void CompileOption::parseAllOptions(map<string, string> &options)
 {
     string val;
 
@@ -166,8 +165,8 @@ CompileOption::parse(const char* argv[], int argc)
             // source files
             string sourceFile = argv[index];
             index++;
-            unsigned found = sourceFile.find_last_of(".");
-            if (found != string::npos) {
+            unsigned int found = sourceFile.find_last_of(".");
+            if (found != (unsigned int)string::npos) {
                 string extension = sourceFile.substr(found);
                 if (extension.empty() || extension != ".tl")
                     return false;
